requirements.txt
# Minimum Django and REST framework version
Django>=1.8
djangorestframework>=3.1

# Test requirements
pytest-django==3.0.0
pytest==3.0.0
pytest-cov==1.6
flake8==2.2.2

# wheel for PyPI installs
wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.11.1
